refactor: Rename `lib/errorHandler.js`
