+++ lib/error-handler.js
