+++ lib/error-handler.test.js
@@ -9,7 +9,7 @@ describe('errorHandler', () => {
   let errorHandler;
   before(() => {
     parseDeploymentData = sinon.stub().resolves({ save: async () => ({ dashboardUrl: 'URL' }) });
-    errorHandler = proxyquire('./errorHandler', {
+    errorHandler = proxyquire('./error-handler', {
       './deployment': { parseDeploymentData },
     });
   });
