+++ lib/plugin.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const { ServerlessSDK } = require('@serverless/platform-client');
 const log = require('./log');
-const errorHandler = require('./errorHandler');
+const errorHandler = require('./error-handler');
 const logsCollection = require('./logsCollection');
 const login = require('./login');
 const logout = require('./logout');
